Get equilibrium age structure for simulation via n iterations

Get equilibrium age structure for simulation via n iterations

Usage

Equilibrium_PopSim_Iter(
  n_regions = n_regions,
  n_sexes = n_sexes,
  n_ages = n_ages,
  init_iter = init_iter,
  NAA,
  M,
  init_F,
  fish_sel,
  r0,
  rec_sexratio,
  movement_matrix
)

Arguments

n_regions

Number of reginos

n_sexes

Number of sexes

n_ages

Number of ages

init_iter

Number of initial iterations to do

NAA

Initial numbers at age array dimensioned by year, region, age, sex, and 1

M

Natural mortality array dimensioned by year, region, age, sex, and 1

init_F

Initial fishing mortality array dimensioned by year, region, fleet, and 1

fish_sel

Fishery selectivity array dimensioned by year, region, age, sex, fleet, and 1

r0

virgin or mean recruitment array dimensioned by year, region, and 1

rec_sexratio

Recruitment sexratio array dimensioned by year, region, sex, and 1

movement_matrix

Movement array, dimensioned by region from, region to, year, age, sex, and 1