Title Truncates data to feedback simulation loop year

Title Truncates data to feedback simulation loop year

Usage

Get_Feedback_Data(
  sim_env,
  sim_list,
  y,
  sim,
  skeleton_data,
  skeleton_parameters,
  skeleton_mapping
)

Arguments

sim_env

Simulation environment defined using Simulate_Pop_Feedback

sim_list

Simulation list that defines the operating model options

y

y index used for annual cycle

sim

sim index used for annual cycle

skeleton_data

Skeleton data list for estimation method (SPoCK) where all dimensions are defined using the total number of simulation years

skeleton_parameters

Skeleton parameters list for estimation method (SPoCK) where all dimensions are defined using the total number of simulation years

skeleton_mapping

Skeleton mapping list for estimation method (SPoCK) where all dimensions are defined using the total number of simulation years

Value

Returns data, parameters, and mapping list truncated to match the feedback year