Wrapper function to get reference points

Wrapper function to get reference points

Usage

Get_Reference_Points(
  data,
  rep,
  SPR_x = NULL,
  t_spwn = 0,
  sex_ratio_f = 0.5,
  calc_rec_st_yr = 1,
  rec_age = 1,
  type,
  what
)

Arguments

data

Data list from RTMB

rep

Report list from RTMB

SPR_x

SPR percentage to target

t_spwn

specified mortality time up until spawning

sex_ratio_f

Sex ratio for females used to compute biological reference points

calc_rec_st_yr

The first year in which mean recruitment is computed for

rec_age

Actual recruitment age value

type

Whether this is a "single_region" reference point or "multi_region"

what

What kind of reference point to use:

SPR

Spawning Potential Ratio for a single region model

independent_SPR

Spawning Potential Ratio for a multi region model, without movement

global_SPR

Global Spawning Potential Ratio for a multi region model, with movement

BH_MSY

MSY reference points derived from a Beverton-Holt for a single region model

independent_BH_MSY

MSY reference points derived from a Beverton-Holt, without movement

global_BH_MSY

MSY reference points derived from a Beverton-Holt, with movement

Value

A list object of fishing and biological reference points

Examples

if (FALSE) { # \dontrun{
f_40 <- Get_Reference_Points(data = data,
rep = rep,
SPR_x = 0.4,
t_spwn = 0,
type = "single_region",
what = "SPR")
f_40$F_x # estimated reference point
} # }