Setup biological inputs for estimation model

Setup biological inputs for estimation model

Setup_Mod_Biologicals(
  input_list,
  WAA,
  MatAA,
  AgeingError = NULL,
  Use_M_prior = 0,
  M_prior = NA,
  fit_lengths = 0,
  SizeAgeTrans = NA,
  M_spec = NULL,
  ...
)

Arguments

input_list

List containing a data list, parameter list, and map list

WAA

Inputs of weight at age dimensioned by n_region, n_year, n_ages, n_sexes

MatAA

Inputs of maturity at age dimensioned by n_region, n_year, n_ages, n_sexes

AgeingError

Inputs of ageing error matrix dimensioned by n_ages, n_ages. Default behavior uses an identity matrix (no ageing error)

Use_M_prior

Whether or not to use natural mortality prior, == 0 don't use, == 1 use

M_prior

Vector of natural mortality priors with the first element representing the mean in normal space, and the second element representing the sd.

fit_lengths

Whether or not to fit length data, == 0 dont fit, == 1 fit

SizeAgeTrans

Size age transition matrix dimensioned by n_regions, n_years, n_lens, n_ages, n_sexes

M_spec

Character specifying options for how to estimate natural mortality. Default is NULL such that it is estimated for 2 sexes or only estimated for a single sex if it is a single sex model. Other options include "est_shared_s" which estimates the same natural mortality rate if n_sexes == 2. The other option is "fix" which fixes all natural mortality parameters.

...

Parameter starting values for ln_M and M_offset if not using default