Setup fishery selectivity and catchability specifications

Setup fishery selectivity and catchability specifications

Setup_Mod_Fishsel_and_Q(
  input_list,
  cont_tv_fish_sel,
  fish_sel_blocks,
  fish_sel_model,
  Use_fish_q_prior = 0,
  fish_q_prior = NA,
  fish_q_blocks,
  fishsel_pe_pars_spec = NULL,
  fish_fixed_sel_pars_spec = NULL,
  fish_q_spec = NULL,
  fish_sel_devs_spec = NULL,
  semipar_age_block_spec = NULL,
  corr_opt_semipar = NULL,
  ...
)

Arguments

input_list

List containing a data list, parameter list, and map list

cont_tv_fish_sel

Specificaiton for continuous time-varying selectivity, character vector dimensioned by n_fish_fleets, where the character is time variation type, _, Fleet, fleet number. time variation types include (none, iid, rw, 3dmarg, 3dcond, 2dar1), and so if we were to specify iid for fleet 1, this would be iid_Fleet_1.

fish_sel_blocks

Specification for fishery selectivity blocks as unique numbers for a given region and fleet, array dimensioned by n_regions, n_years, n_fish_fleets

fish_sel_model

Specification for fishery selectivity model for a given region and fleet, array dimensioned by n_regions, n_years, n_fish_fleets, == 0 a50, k, logistic, == 1 gamma dome shaped, == 2, exponential, == 3 a50, a95 logistic

Use_fish_q_prior

Integer specifying whether to use fishery q prior or not (0 dont use) (1 use)

fish_q_prior

Fishery q priors in normal space, dimensioned by region, block, fishery fleet, and 2 (mean, and sd in the 4 dimension of array)

fish_q_blocks

Specification for fishery catchability blocks as unique numbers for a given region and fleet, array dimensioned by n_regions, n_years, n_fish_fleets

fishsel_pe_pars_spec

Specification for fishery selectivity process error parameters. If cont_tv_fish_sel is = 0, then this is all fixed and not estimated. Otherwise, the options are: est_all, which estiamtes all parameters, est_shared_r, which estiamtes parameters shared across regions, est_shared_s, which estiamtes parameters shared across sexes, and est_shared_r_s, which estimates these paraemters shared across regions and sexes

fish_fixed_sel_pars_spec

Specification for fishery selectivity fixed effects parameters. Options are est_all, which estiamtes all parameters, est_shared_r, which estiamtes parameters shared across regions, est_shared_s, which estiamtes parameters shared across sexes, and est_shared_r_s, which estimates these paraemters shared across regions and sexes

fish_q_spec

Specification for fishery catchability. Options are est_all, which estiamtes all parameters across regions, est_shared_r, which estimates parameters shared across regions.

fish_sel_devs_spec

Specificaiton for selectivity process error dviations. Options are est_all, which estimates all deviations, est_shared_r, which shares them across regions, est_shared_s, which shares them across sexes, est_shared_r_s, which shares them across regions and sexes, and est_shared_a, which shares them across age blocks (need to define a number in semipar_age_block_spec), est_shared_r_a, which shares them across regions and age blocks (need to define semipar_age_block_spec), and est_shared_r_a_s, which shares them across regions, ages, and sexes

semipar_age_block_spec

Number that needs to be specified if est_shared_a variants are specified. Number represents the number of ages to share deviations for spaced evenly.

corr_opt_semipar

Only used if cont_tv_sel is 3,4,5. Allows users to turn off estimation of certain correlation parameters ot be at 0. Options include corr_zero_y, which turns year correlations to 0, corr_zero_a which turns age correaltions to 0, corr_zero_y_a which turns year age correaltions to 0. These options can be used for cont_tv_sel 3,4,5. Additional options include corr_zero_c, which turns cohort correaltions to 0, corr_zero_y_c, which turns cohort and year correaltions to 0, corr_zero_a_c which turns age and cohort correaltions to 0, as well as corr_zero_y_a_c, which turns all correlations to 0, and effectively collapses to iid. These latter options are only available for cont_tv_sel 3,4.

...

Additional arguments specifying starting values for fishery selectivity and catchability parameters (fishsel_pe_pars, ln_fishsel_devs, ln_fish_fixed_sel_pars, ln_fish_q)