Title

Title

Setup_Mod_Weighting(
  input_list,
  sablefish_ADMB,
  likelihoods,
  Wt_Catch = 1,
  Wt_FishIdx = 1,
  Wt_SrvIdx = 1,
  Wt_Rec = 1,
  Wt_F = 1,
  Wt_FishAgeComps,
  Wt_SrvAgeComps,
  Wt_FishLenComps,
  Wt_SrvLenComps
)

Arguments

input_list

List containing a data list, parameter list, and map list

sablefish_ADMB

Numeric indicating whether to mimic calculations for the sablefish ADMB model (1) or not (0)

likelihoods

Numeric indicating whether to use ADMB likelihoods (0) or TMB likelihoods (1)

Wt_Catch

Weight (lambda) applied to the overall catch dataset

Wt_FishIdx

Weight (lambda) applied to the overall fishery index dataset

Wt_SrvIdx

Weight (lambda) applied to the overall survey index dataset

Wt_Rec

Weight (lambda) applied to recruitment penalty

Wt_F

Weight (lambda) applied to fishing mortality penalty

Wt_FishAgeComps

Weight (lambda) applied to fishery age compositions

Wt_SrvAgeComps

Weight (lambda) applied to survey age compositions

Wt_FishLenComps

Weight (lambda) applied to fishery length compositions

Wt_SrvLenComps

Weight (lambda) applied to survey length compositions