Setup values and dimensions of fishing mortality

Setup values and dimensions of fishing mortality

Usage

Setup_Sim_FishMort(
  n_sims = n_sims,
  n_yrs = n_yrs,
  n_regions = n_regions,
  n_fish_fleets = n_fish_fleets,
  sigmaC,
  init_F_vals,
  Fmort_pattern,
  Fmort_start,
  Fmort_fct,
  proc_error,
  proc_error_sd
)

Arguments

n_sims

Number of simulations

n_yrs

Number of years

n_regions

Number of regions

n_fish_fleets

Number of fleets

sigmaC

Observation error for catch

init_F_vals

Initial F values dimensioend by region and fleet

Fmort_pattern

Fishing mortality pattern as a matrix dimensioned by region and fleet (constant, linear, one-way, two-way)

Fmort_start

Fishing mortality start values as a matrix dimensioned by region and fleet

Fmort_fct

Fishing mortality values for factor increases or decreases as a matrix dimensioned by region and fleet

proc_error

If we want to add process error to fishing mortality

proc_error_sd

value of logrnomal sd for process error to fishing mortality