Setup fishery selectivity

Setup fishery selectivity

Usage

Setup_Sim_FishSel(
  n_sims = n_sims,
  n_yrs = n_yrs,
  n_regions = n_regions,
  n_ages = n_ages,
  n_sexes = n_sexes,
  n_fish_fleets = n_fish_fleets,
  sel_model,
  fixed_fish_sel_pars
)

Arguments

n_sims

Number of simulations

n_yrs

Number of years

n_regions

Number of regions

n_ages

Number of ages

n_sexes

Number of sexes

n_fish_fleets

Number of fishery fleets

sel_model

Fishery selectivity model dimensioned by region and fleet. Options include: logistic

fixed_fish_sel_pars

Fixed parameters of fishery selectivity, dimensioned by region, sex, fishery fleet, and the max number of parameters needed for for a defined fishery selectivity functional form out of all defined functional forms for the fishery