Sablefish data for single region case study

A dataset containing the necessary elements for the Alaska sablefish case study.

Usage

sgl_rg_sable_data

Format

A list with multiple components needed for the single region sablefish model

Source

2024 Federal Alaska Sablefish Assessment