Setup model movement processes

Setup model movement processes

Setup_Mod_Movement(
  input_list,
  do_recruits_move = 0,
  use_fixed_movement = 0,
  Fixed_Movement = NA,
  Use_Movement_Prior = 0,
  Movement_prior = NULL,
  Movement_ageblk_spec = NULL,
  Movement_yearblk_spec = NULL,
  Movement_sexblk_spec = NULL,
  ...
)

Arguments

input_list

List containing a data list, parameter list, and map list

do_recruits_move

Whether recruits move, 1 (they move), 0 (they don't move), default is 0

use_fixed_movement

Whether to use a fixed movement matrix, == 0 don't use, == 1 use

Fixed_Movement

Input for a fixed movement matrix dimensioned by n_regions, n_regions, n_years, n_ages, n_sexes

Use_Movement_Prior

Whether or not to use movement priors == 0 don't use, == 1 use

Movement_prior

Movement prior values, can be dimension as a vector of (results in a constant prior across all regions from, years, ages, and sexes), or as an array by n_regions, n_regions, n_years, n_ages, n_sexes

Movement_ageblk_spec

Either a character string specifiying "constant" age movement, or a ist object specifying which ages to group together and block (i.e., which ages to share parameters for). Each element in the list should be a vector specifying the range of ages for which to share or block parameters for. For example, list(c(1:6), c(7:10), c(11:n_ages)) has 3 age blocks, where ages 1-6 have the same paraemters, 7:10 have the same parameters (but diffeerent from 1:6), and 11:n_ages haave the same parameters (but different from 1:6 and 7:10). The default behavior is to estimate unique movement parmaeters for all ages. To estimate age-invariant movement this list would be specified as: list(c(1:n_ages)).

Movement_yearblk_spec

Either a character string specifiying "constant" year movement, list object specifying which years to group together and block (i.e., which years to share parameters for). Each element in the list should be a vector specifying the range of years for which to share or block parameters for. For example, list(c(1:6), c(7:10), c(11:n_years)) has 3 year blocks, where years 1-6 have the same paraemters, 7:10 have the same parameters (but diffeerent from 1:6), and 11:n_years haave the same parameters (but different from 1:6 and 7:10). The default behavior is to estimate unique movement parameters for all years To estimate time-invariant movement this list would be specified as: list(c(1:n_years)).

Movement_sexblk_spec

Either a character string specifiying "constant" sex movement, list object specifying which sexes to group together and block (i.e., which sexes to share parameters for). Each element in the list should be a vector specifying the range of sexes for which to share or block parameters for. For example, list(1, 2) has 2 sex blocks, where each sex has unique parameters The default behavior is to estimate unique movement parameters for all sexes. To estimate sex-invariant movement this list would be specified as: list(c(1:n_sexes)).

...

Additional inputs for starting values for movement parameters (move_pars)