Setup model objects for specifying recruitment module and associated processes

Setup model objects for specifying recruitment module and associated processes

Usage

Setup_Mod_Rec(
  input_list,
  rec_model,
  rec_dd = NULL,
  rec_lag = 1,
  Use_h_prior = 0,
  h_mu = NA,
  h_sd = NA,
  do_rec_bias_ramp = 0,
  bias_year = NA,
  sigmaR_switch = 1,
  dont_est_recdev_last = 0,
  sexratio = 1,
  init_age_strc = 0,
  init_F_prop = 0,
  sigmaR_spec = NULL,
  InitDevs_spec = NULL,
  RecDevs_spec = NULL,
  h_spec = NULL,
  t_spawn = 0,
  ...
)

Arguments

input_list

List containing a data list, parameter list, and map list

rec_model

recruitment model, options are mean_rec, bh_rec

rec_dd

Recruitment density dependence, Options are "local", "global", or NULL

rec_lag

duration of recruitment lag with SSB

Use_h_prior

whether or not to use a steepness prior == 0 (don't use), == 1 (use)

h_mu

vector of steepness mean if priors are used (dimensioned by n_region)

h_sd

vector of steepness sd if priors are used (dimensioned by n_region)

do_rec_bias_ramp

whether or not to do recruitment bias ramp == 0 (don't use), == 1(use)

bias_year

vector of years in which the bias ramp is invoked and when it changes. This should have a total of 4 values, the first describing when the period in which no bias correction is applied ends, the second describing the period over which the ascending limb ends, the third describing the period in which the full bias correction is last applied, and the fourth describing the period where no bias ramp is applied, following the period of full bias correction.

sigmaR_switch

year in which sigmaR switches from an early value to late (set to 0 if no switching occurs)

dont_est_recdev_last

Numeric indicating the last x years to not estimate rec devs for. Default is 0.

sexratio

vector of recruitment sex ratio dimensioned by n_sexes

init_age_strc

whether initial age structure is initialized via == 0 (iteration) or via == 1 (geometric series)

init_F_prop

Initial F proportion relative to a mean F value for initializing age structure

sigmaR_spec

Specification for sigmaR. Default is NULL such that it is estimated for both early and late periods. Other options include "est_shared" which estiamtes sigmaR but shares the same value between early and late period and "fix" which fixes both values.

InitDevs_spec

Specification for initial age dviations. Default is NULL such that is its estimated for all ages and regions. Other options include "est_shared_r" which estimates deviations for all ages but shares them across regions, or "fix" which fixeds all values.

RecDevs_spec

Specification for recruitment deviations. Default is NULL such that it is estimated for all regions and years. Other options include "est_shared_r" which estimates deviations for all years, but shares them across regions (global recruitment deviaitons), or "fix" which fixes all values.

h_spec

Specification for steepness. Default is NULL, such that it is estimated for all reigons if rec_model == 1 (Beverton-Holt). Other options include "est_shared_r" which estimates h but shares across regions, or "fix" which fixes all steepness values. If rec_model == 0, h is fixed.

t_spawn

Spawn timing fraction

...

Additional arguments for starting values for recruitment parameters (ln_global_R0, Rec_prop, h, ln_InitDevs, ln_RecDevs, ln_sigmaR)