Function to do likelihood profile on a parameter

Function to do likelihood profile on a parameter

do_likelihood_profile(
  data,
  parameters,
  mapping,
  random = NULL,
  what,
  idx = NULL,
  min_val,
  max_val,
  inc = 0.05
)

Arguments

data

data list from model

parameters

parameter list from model

mapping

mapping list from model

random

character vector of random effects to estimate

what

parameter name we want to profile

idx

Index for an parameter array, pointing to the value we want to map off

min_val

minimum value of profile

max_val

maximum value of profile

inc

increment value between min and max value

Value

Returns a list of likelihood profiled values for each data component with their respective dimensions (e.g., likelihood profiles by fleet, region, year, etc.) as well likelihood profiles for each data component, aggregated across all their respective dimensions.